[export] Add ability to run eagerly on UnflattenedModule
Summary: Added the contextmanager, `_disable_interpreter`, which is meant to put around a call to unflatten. This will generate an UnflattendModule and sub-InterpreterModules which will not use torch.fx.Interpreter to run eagerly. This seems to improve the performance: https://fb.workplace.com/groups/1075192433118967/posts/1473590629945810/?comment_id=1473621763276030

Test Plan: CI

Differential Revision: D60426372